% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbrdollars.R
\name{abbrdollars}
\alias{abbrdollars}
\title{Abbreviate Dollars}
\usage{
abbrdollars(
  x,
  thousandabbr = "K",
  millionabbr = "M",
  billionabbr = "B",
  currency = "$"
)
}
\arguments{
\item{x}{a numeric vector of dollar values}

\item{thousandabbr}{a single character string to be used as the abbreviation for thousands}

\item{millionabbr}{a single character string to be used as the abbreviation for thousands}

\item{billionabbr}{a single character string to be used as the abbreviation for thousands}

\item{currency}{the currency symbol to be used. Prepended to each vaue of x}
}
\value{
a character vector of abbreviated dollar values. Values are rounded to 3 significant digits using `floor()`.
}
\description{
format dollars abbreviate
}
