% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistency.R
\name{consistency}
\alias{consistency}
\title{Calculate the consistency rate for constituents (the percentage of time they have given during a defined range of time)}
\usage{
consistency(data, id_col, fy_col, threshold = 1, alt_year = NA,
  max_year = 0)
}
\arguments{
\item{data}{a data frame}

\item{id_col}{a column from the data frame containing constituent IDs - column name should be unquoted}

\item{fy_col}{a column from the data frame containing fiscal years for gifts - column name should be unquoted (use the fy() function if fiscal years are not already in the data)}

\item{threshold}{a numeric value for the minimum active years to consider - for example, if someone is a first time donor they will have 100% consistency to only include those that could have given at least 3 fiscal years, change threshold to 3}

\item{alt_year}{a numeric value or a quoted column name from the data frame containing which fiscal years should be used for the lower bound of the active years range. By default, the minimum fiscal year for gifts will be used however graduation year could also be used or just any given fiscal year}

\item{max_year}{a numeric value for the upper bound of the active years. By default, the maximum fiscal year for gifts will be used however any fiscal year can be used to override this such as the current fiscal year}
}
\value{
a summary data frame with the unique ID, first and last year of the active range, the defined active giving years, the number of fiscal years during the active giving years when a gift was a received, the consistency rate of fiscal years with a gift over active giving range
}
\description{
calculates consistency a rate of how often gifts have been made during a defined time period
}
