% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contact_data}
\alias{contact_data}
\title{Sample Contact Reports}
\format{A data frame with 196 rows and 7 variables:
\describe{
  \item{Staff Name}{the name of the development officer}
  \item{Method}{the contact method: Email, Letter, Phone}
  \item{Date}{the date of the contact}
  \item{Summary}{descripitive text from the contact}
  \item{Substantive}{flag to indicate if contact was substantive}
  \item{Donor}{the name of the donor or prospect}
  \item{Outcome}{flag to indicate the outcome of the contact}
}}
\source{
\url{https://www.kaggle.com/michaelpawlus/fundraising-data}
}
\usage{
contact_data
}
\description{
A dataset containing 196 contact reports. This dataset is
included with the book Data Science for Fundraising and
is used with permission in this package
}
\keyword{datasets}
