% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency_to_numeric.R
\name{currency_to_numeric}
\alias{currency_to_numeric}
\title{Currency to Numeric}
\usage{
currency_to_numeric(x)
}
\arguments{
\item{x}{a vector of dollar amounts}
}
\value{
a numeric vector
}
\description{
coerce dollar amounts to numeric values. Removes "$" and "," and coerces to numeric.
}
