% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutgiving.R
\name{cutgiving}
\alias{cutgiving}
\title{Cut Giving}
\usage{
cutgiving(x)
}
\arguments{
\item{x}{a numeric vector of gifts or gift bins}
}
\value{
a factor vector with bins based on the values in x
}
\description{
a version of cut designed to cut a vector of gifts (or gift totals) into useful bins
}
\details{
`cutgiving()` categorizes x into a series of bins, much like a histogram would, but uses `asinh()` to log transform the data first, working on the assumption that giving will be highly skewed.
}
