% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{donor_data}
\alias{donor_data}
\title{Donor Data}
\format{A data frame with 3451 rows and 23 variables:
\describe{
  \item{ID}{the unique identifier for each constituent}
  \item{ZIPCODE}{the ZIP code for the constituent}
  \item{AGE}{the age of the constituent}
  \item{MARITAL_STATUS}{the marital status of the constituent}
  \item{GENDER}{the gender of the constituent}
  \item{MEMBERSHIP_IND}{a binary flag indicating if a constituent is a radio member}
  \item{ALUMNUS_IND}{a binary flag indicating if a constituent is an alumnus}
  \item{PARENT_IND}{a binary flag indicating if a constituent is a parent}
  \item{HAS_INVOLVEMENT_IND}{a binary flag indicating if a constituent has an involvement code}
  \item{WEALTH_RATING}{a descriptive factor with the results of a wealth screen}
  \item{DEGREE_LEVEL}{the degree level of the constituent if they are an alumnus}
  \item{PREF_ADDRESS_TYPE}{the preferred address type for the constituent}
  \item{EMAIL_PRESENT_IND}{a binary flag indicating if an email address is present}
  \item{CON_YEARS}{consecutive giving years}
  \item{PrevFYGiving}{the fiscal year giving for the previous fiscal year}
  \item{PrevFY1Giving}{the fiscal year giving for the year before the previous fiscal year}
  \item{PrevFY2Giving}{the fiscal year giving for two years before the previous fiscal year}
  \item{PrevFY3Giving}{the fiscal year giving for three years before the previous fiscal year}
  \item{PrevFY4Giving}{the fiscal year giving for four years before the previous fiscal year}
  \item{CurrFYGiving}{the fiscal year giving for the current fiscal year}
  \item{TotalGiving}{the total giving for the constituent}
  \item{DONOR_IND}{a binary flag indicating if a constituent is a donor}
  \item{BIRTH_DATE}{the birth date for the constituent}
}}
\source{
\url{https://www.kaggle.com/michaelpawlus/fundraising-data}
}
\usage{
donor_data
}
\description{
A dataset containing 3451 constituent rows and 23 features.
This dataset is a random 10% slice of the dataset included
with the book Data Science for Fundraising and is used with
permission in this package. The dataset was sampled to keep
the object size under 1MB
}
\keyword{datasets}
