% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fy_quarters.R
\name{fiscal_months}
\alias{fiscal_months}
\title{Create a lookup table that calculates fiscal months based on the start of the fiscal calendar}
\usage{
fiscal_months(fy_start_month = 7)
}
\arguments{
\item{fy_start_month}{an integer between 1 ans 12 indicating the first month of the fiscal year}
}
\value{
a tibble with two columns the calendar month and the fiscal month based on the month that starts the fiscal year
}
\description{
an intermediate function - uses the calendar month that starts the fiscal year to create a table with fiscal months
}
