% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lysy.R
\name{flag_lapsed}
\alias{flag_lapsed}
\title{Identify donors who gave last year or some year other than last but are not current donors}
\usage{
flag_lapsed(x, last_gift_date, fy_start_month = 7)
}
\arguments{
\item{x}{a data set}

\item{last_gift_date}{a column with the last gift date}

\item{fy_start_month}{the number of the fiscal start month}
}
\value{
a tibble with a column identifying lapsed donors
}
\description{
create a column identifying lapsed donors
}
