% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fy.R
\name{fy}
\alias{fy}
\title{Calculate the MU fiscal year}
\usage{
fy(x)
}
\arguments{
\item{x}{a date-time object}
}
\value{
a numeric value of fiscal years
}
\description{
calculates the fiscal year based on an existing date
}
