% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fy2.R
\name{fy2}
\alias{fy2}
\title{Calculate the fiscal year based on any starting month}
\usage{
fy2(date, fy_start_month = 7)
}
\arguments{
\item{date}{a date-time object}

\item{fy_start_month}{a number represeting the starting month for the fiscal year (defaults to 7 -- "July")}
}
\value{
a numeric value of fiscal years
}
\description{
calculates the fiscal year based on an existing date
}
