% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fy_quarters.R
\name{fy_quarters}
\alias{fy_quarters}
\title{Append a data frame with the gift month, fiscal month and fiscal quarter}
\usage{
fy_quarters(x, gift_date_col, fy_start_month = 7)
}
\arguments{
\item{x}{a data frame}

\item{gift_date_col}{a column with gift dates}

\item{fy_start_month}{an integer between 1 ans 12 indicating the first month of the fiscal year}
}
\value{
the data frame provided by the user with three additional columns: gift_month, fiscal_month and fiscal_quarter
}
\description{
Uses the start month for the fiscal year to calculate the fiscal month and fiscal quarter
}
