% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fydaynum.R
\name{fydaynum}
\alias{fydaynum}
\title{Fiscal Year Day Number}
\usage{
fydaynum(x = Sys.Date(), ytd = TRUE)
}
\arguments{
\item{x}{a date to convert to a numeric day of the fiscal year. Note that x should be of class "Date".}

\item{ytd}{a logical vector. If true, only values for dates that are year-to-date will be returned.}
}
\value{
a numeric value describing what day of the fiscal year x is.
}
\description{
Fiscal Year Day Number
}
