% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_screening.R
\name{get_block_id_table}
\alias{get_block_id_table}
\title{Iterate through get_block_ids to create a table of block ids - a list for now but a table soon}
\usage{
get_block_id_table(result)
}
\arguments{
\item{result}{a data frane output as a result of the get_lat_lon function containing a value for longitude and latitude labelled lon and lat respectively}
}
\value{
a data frame with block ids
}
\description{
creates a data frame with block ids generated from longitude and latitude values
}
