% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_screening.R
\name{get_block_ids}
\alias{get_block_ids}
\title{Get a block id for every longitude and latitude value - a helper function that is iterated through to make a table}
\usage{
get_block_ids(lat, lon)
}
\arguments{
\item{lat}{a value for latitude}

\item{lon}{a value for longitude}
}
\value{
block id value or values
}
\description{
pulls a block id values for a given set of coordinates
}
