% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_screening.R
\name{get_lat_lon}
\alias{get_lat_lon}
\title{Get the latitide and longitude values based on address data}
\usage{
get_lat_lon(x, address, name)
}
\arguments{
\item{x}{a data frame}

\item{address}{a column from the data frame containing a concatenated address line}

\item{name}{a column from the data frame containing a name field with a full name}
}
\value{
a data frame with latitide and longitude data for every address
}
\description{
creates a data frame with latitude and longitude values based on address data
}
