% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longest_consecutive.R
\name{longest_consecutive}
\alias{longest_consecutive}
\title{Calculate the longest consecutive giving streak}
\usage{
longest_consecutive(data, ids, fys, fy_range = NA, mode = "append")
}
\arguments{
\item{data}{a data frame with donor ids and the fiscal year of gifts}

\item{ids}{an unquoted column name that contains the constituent IDs}

\item{fys}{an unquote column name that contains the fiscal year for gifts. If not present, run fy() or fy2().}

\item{fy_range}{the maximum consecutive giving streak to try to find. By default, this uses the maximum range given the fiscal years provided.}

\item{mode}{options for the output mode. By default, "append" creates a new data frame with all provided columns and the consecutive giving years column. There are two other options. "just_consecutive" will provide a table of IDs and values representing the longest consecutive giving streak. "summary" provides a count of donors for every consecutive giving value.}
}
\value{
a data frame of one out three possible options controlled by the value passed to the mode argument
}
\description{
calculates the longest consecutive giving streak given giving details for constituents containing a fiscal year field
}
