% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_screening.R
\name{make_full_name}
\alias{make_full_name}
\title{Create a full name column using the first and last names}
\usage{
make_full_name(x, first_name_col, last_name_col)
}
\arguments{
\item{x}{a data frame}

\item{first_name_col}{a column from the data frame containing the first name}

\item{last_name_col}{a column from the data frame containing the last name}
}
\value{
a given data frame with an appended column containthing the full name - in this case: first name, space, last name
}
\description{
creates a full name column - a convenience helper function for acs screening
}
