% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{network_data}
\alias{network_data}
\title{Network Dataset}
\format{A data frame with 30 rows and 2 variables:
\describe{
  \item{from}{the starting point for the relationship}
  \item{to}{the ending point for the relationship}
}}
\source{
\url{https://www.kaggle.com/michaelpawlus/fundraising-data}
}
\usage{
network_data
}
\description{
A dataset containing 30 links between entities for exploring network
relationships. This dataset is included with the book Data Science
for Fundraising and is used with permission in this package.
}
\keyword{datasets}
