% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips.R
\name{parse_fips}
\alias{parse_fips}
\title{Parsing FIPS codes}
\usage{
parse_fips(fips)
}
\arguments{
\item{fips}{FIPS code, as a 5 to 15 character string, containing only numeric characters}
}
\value{
Named list of FIPS codes.
}
\description{
A single FIPS code concatenates multiple points of information into a single
string. This function parses the 2-digit state, 3-digit county, 5-digit
tract, block group, and 4-digit block codes (if present) from a 5-15 character long string.
}
\examples{
parse_fips("090010218022000")

}
