% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_screening.R
\name{query_acs}
\alias{query_acs}
\title{Obtain 5-Year American Community Survey Estimates}
\usage{
query_acs(var = c("B19013_001E", "B25077_001E"), year = NULL, state,
  county = NULL, tract = NULL, blkgrp = NULL, key = NULL)
}
\arguments{
\item{var}{Variables to query from the ACS. For a list of the available
variables and codes (for the 2017 ACS), see the
\href{https://api.census.gov/data/2017/acs/acs5/variables.html}{Official
Documentation}. Defaults to Median Household Income (B19013_00E) and Median
Home Value (Owner-Occupied Units) (B25077_001E). Supports groups.}

\item{year}{Four-digit year. Defaults to the most recent data, for 2017.}

\item{state}{Two-digit state FIPS code. Alternatively, \code{"us"} for
national-level statistics. Supports wildcard string (\code{"*"}).}

\item{county}{Three-digit county FIPS code. Supports wildcard string
(\code{"*"}).}

\item{tract}{Five-digit census tract FIPS code. Supports wildcard string
(\code{"*"}).}

\item{blkgrp}{One-digit blog group FIPS code.}

\item{key}{(optional) Developer key.}
}
\value{
Tibble of data points and FIPS codes, one line per valid input geography.
}
\description{
The U.S. Census Bureau has published 5-year esimates of
  demographic data since 2009. The data is aggregated at the national, state,
  county, census tract, and census block group levels.

  This function queries the
  \href{https://www.census.gov/data/developers/data-sets/acs-5year.html}{Census
   Bureau API} based on FIPS codes for various geographies. Substituting a
  wildcard character \code{"*"} instead of a FIPS code returns all values
  within the parent geography (i.e. \code{tract = "*"} will return data for
  all tracts within a county).


  The API limits the number of queries for users who lack an API key. A key
  can be obtained \href{https://api.census.gov/data/key_signup.html}{here}.
}
