% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_data}
\alias{small_data}
\title{Small Alumni Dataset}
\format{
A data frame with 2878 rows and 8 variables:
\describe{
  \item{coreid}{a unique id for an alumni record}
  \item{aff_total}{a calculated affinity score - mark of how connected an alum is to the university}
  \item{coresex}{the gender of the alumnus}
  \item{coreprefyr}{the preferred graduation year for the alumnus}
  \item{Degree}{the degree earned by the alumnus}
  \item{Major}{the major for the alumnus}
  \item{fy14}{the total giving for the alumnus during fiscal year 2014}
  \item{donor}{a binary flag of 1 or 0 to indicate if the alumnus is a donor}
}
}
\source{
\url{https://www.kaggle.com/michaelpawlus/fundraising-data}
}
\usage{
small_data
}
\description{
A dataset containing 2878 anonymized alumni records
along with graduate year, degree, major, giving total
and an affinity score. This dataset was created by
Michael Pawlus.
}
\keyword{datasets}
