% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survey_data}
\alias{survey_data}
\title{Survey Responses}
\format{
A data frame with 30 rows and 2 variables:
\describe{
  \item{ID Number}{the unique identifier for each survey taker}
  \item{Survey Response}{a numeric value reflecting a survey response}
}
}
\source{
\url{https://www.kaggle.com/michaelpawlus/fundraising-data}
}
\usage{
survey_data
}
\description{
A dataset containing 1263 survey responses where the survey responses
are expressed as a single numeric value for every respondent. This dataset
in included with the book Cool Data: A how-to guide... and is used with
permission in this package.
}
\keyword{datasets}
