% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fy2.R
\name{fy2}
\alias{fy2}
\title{Calculate the fiscal year based on any starting month}
\usage{
fy2(date = Sys.Date(), fy_start_month = 7)
}
\arguments{
\item{date}{a vector of date or date-times}

\item{fy_start_month}{a number representing the starting month for the fiscal year (defaults to 7, ie. July)}
}
\value{
a numeric vector of fiscal years with the length of the date argument
}
\description{
calculates the fiscal year based on an existing date
}
