% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_screening.R
\name{get_lat_lon}
\alias{get_lat_lon}
\title{Get the latitide and longitude values based on address data}
\usage{
get_lat_lon(name, address)
}
\arguments{
\item{name}{a vector containing the full name for the record as a string most likely in the form of data dollar sign column}

\item{address}{a vector containing the concatenated address as a string most likely in the form of data dollar sign column}
}
\value{
a data frame with latitide and longitude data for every address
}
\description{
creates a data frame with latitude and longitude values based on address data
}
